requirements.txt
torch>=1.8
torchvision
tqdm
Pillow
